# Gather to Slice Fusion