--- a/onnxruntime/core/optimizer/gather_fusion.cc
+++ b/onnxruntime/core/optimizer/gather_fusion.cc
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "core/optimizer/gather_to_split_fusion.h"
+#include "core/optimizer/gather_fusion.h"
 
 #include "core/graph/graph_utils.h"
 #include "core/optimizer/initializer.h"
@@ -175,4 +175,118 @@ Status GatherToSplitFusion::ApplyImpl(Graph& graph, bool& modified, int graph_le
 
   return Status::OK();
 }
+
+/*
+Fuse Range->Gather to Slice.
+*/
+Status GatherToSliceFusion::ApplyImpl(Graph& graph, bool& modified, int graph_level,
+                                      const logging::Logger& logger) const {
+  GraphViewer graph_viewer(graph);
+  const auto& node_topology_list = graph_viewer.GetNodesInTopologicalOrder();
+
+  for (auto node_index : node_topology_list) {
+    auto* p_node = graph.GetNode(node_index);
+    if (p_node == nullptr) continue;  // we removed the node as part of an earlier fusion
+    Node& node = *p_node;
+
+    ORT_RETURN_IF_ERROR(Recurse(node, modified, graph_level, logger));
+
+    if (!graph_utils::IsSupportedOptypeVersionAndDomain(node, "Range", {1, 11}) ||
+        !graph_utils::IsSupportedProvider(node, GetCompatibleExecutionProviders()) || node.GetOutputEdgesCount() != 1) {
+      continue;
+    }
+
+    Node& gather_node = *graph.GetNode(node.OutputNodesBegin()->Index());
+    if (!graph_utils::IsSupportedOptypeVersionAndDomain(gather_node, "Gather", {1, 11, 13}) ||
+        !graph_utils::IsSupportedProvider(gather_node, GetCompatibleExecutionProviders())) {
+      continue;
+    }
+
+    // Range's output is Gather's input[1].
+    if (node.MutableOutputDefs()[0] != gather_node.MutableInputDefs()[1]) {
+      continue;
+    }
+
+    InlinedVector<std::reference_wrapper<Node>> nodes_to_fuse{node, gather_node};
+
+    auto& range_input_defs = node.MutableInputDefs();
+    ORT_ENFORCE(range_input_defs.size() == 3);
+    // Range's inputs are scalar, need unsqueeze to 1-D tensors.
+    ONNX_NAMESPACE::TypeProto unsqueeze_output_type;
+    const ONNX_NAMESPACE::TensorProto_DataType element_type = static_cast<ONNX_NAMESPACE::TensorProto_DataType>(
+        node.MutableOutputDefs()[0]->TypeAsProto()->tensor_type().elem_type());
+    unsqueeze_output_type.mutable_tensor_type()->set_elem_type(element_type);
+    unsqueeze_output_type.mutable_tensor_type()->mutable_shape()->add_dim()->set_dim_value(1LL);
+
+    InlinedVector<NodeArg*> unsqueeze_outputs;
+    for (size_t i = 0; i < range_input_defs.size(); ++i) {
+      unsqueeze_outputs.emplace_back(&graph.GetOrCreateNodeArg(
+          graph.GenerateNodeArgName("unsqueeze_output_" + std::to_string(i)), &unsqueeze_output_type));
+    }
+
+    // Unsqueeze before and after OpSet-13 have different schemas.
+    int onnx_opset_version = -1;
+    if (graph.DomainToVersionMap().find(kOnnxDomain) != graph.DomainToVersionMap().end()) {
+      onnx_opset_version = graph.DomainToVersionMap().at(kOnnxDomain);
+    }
+
+    if (onnx_opset_version < 13) {
+      for (size_t i = 0; i < range_input_defs.size(); ++i) {
+        Node& unsqueeze_node =
+            graph.AddNode(graph.GenerateNodeName("Unsqueeze_" + std::to_string(i)), "Unsqueeze",
+                          "Unsqueeze for Fused Gather nodes", {range_input_defs[i]}, {unsqueeze_outputs[i]});
+        unsqueeze_node.AddAttribute("axes", std::vector<int64_t>{static_cast<int64_t>(0)});
+        unsqueeze_node.SetExecutionProviderType(node.GetExecutionProviderType());
+      }
+    } else {
+      ONNX_NAMESPACE::TensorProto unsqueeze_axes_initializer_proto;
+      unsqueeze_axes_initializer_proto.set_name(graph.GenerateNodeName("UnsqueezeAxesInitializer"));
+      unsqueeze_axes_initializer_proto.add_dims(static_cast<int64_t>(1));
+      unsqueeze_axes_initializer_proto.set_data_type(ONNX_NAMESPACE::TensorProto_DataType_INT64);
+      unsqueeze_axes_initializer_proto.add_int64_data(static_cast<int64_t>(0));
+      NodeArg* unsqueeze_axes_arg = &graph_utils::AddInitializer(graph, unsqueeze_axes_initializer_proto);
+
+      for (size_t i = 0; i < range_input_defs.size(); ++i) {
+        Node& unsqueeze_node = graph.AddNode(graph.GenerateNodeName("Unsqueeze_" + std::to_string(i)), "Unsqueeze",
+                                             "Unsqueeze for Fused Gather nodes",
+                                             {range_input_defs[i], unsqueeze_axes_arg}, {unsqueeze_outputs[i]});
+        unsqueeze_node.SetExecutionProviderType(node.GetExecutionProviderType());
+      }
+    }
+
+    int64_t axis = 0;  // Default value.
+    auto& attrs = gather_node.GetAttributes();
+    if (attrs.find("axis") != attrs.end()) {
+      auto& axis_attr = attrs.at("axis");
+      if (utils::HasInt(axis_attr)) axis = axis_attr.i();
+    }
+
+    ONNX_NAMESPACE::TensorProto slice_axes_initializer_proto;
+    slice_axes_initializer_proto.set_name(graph.GenerateNodeName("SliceAxesInitializer"));
+    slice_axes_initializer_proto.add_dims(static_cast<int64_t>(1));
+    slice_axes_initializer_proto.set_data_type(element_type);
+    // Tind of Slice can only support int32 and int64.
+    if (element_type == ONNX_NAMESPACE::TensorProto_DataType_INT64) {
+      slice_axes_initializer_proto.add_int64_data(axis);
+    } else {
+      slice_axes_initializer_proto.add_int32_data(static_cast<int32_t>(axis));
+    }
+    NodeArg* slice_axes_arg = &graph_utils::AddInitializer(graph, slice_axes_initializer_proto);
+    Node& slice_node = graph.AddNode(graph.GenerateNodeName("Slice"), "Slice", "Slice for Fused Gather nodes",
+                                     {gather_node.MutableInputDefs()[0], unsqueeze_outputs[0], unsqueeze_outputs[1],
+                                      slice_axes_arg, unsqueeze_outputs[2]},
+                                     {gather_node.MutableOutputDefs()[0]});
+    slice_node.SetExecutionProviderType(gather_node.GetExecutionProviderType());
+
+    for (Node& n : nodes_to_fuse) {
+      graph_utils::RemoveNodeOutputEdges(graph, n);
+      graph.RemoveNode(n.Index());
+    }
+
+    modified = true;
+  }
+
+  return Status::OK();
+}
+
 }  // namespace onnxruntime